Pass `filter_params` to the calendar filtering instead of the request params
