+++ decidim-meetings/app/controllers/decidim/meetings/calendars_controller.rb
@@ -5,13 +5,17 @@ module Decidim
     # Exposes the meeting resources as an .ics file so users can import them
     # to their favorite calendar app
     class CalendarsController < Decidim::Meetings::ApplicationController
+      include FilterResource
+      include Filterable
+      include ComponentFilterable
+
       layout false
       helper_method :meetings
       before_action :set_default_request_format
       skip_around_action :use_organization_time_zone
 
       def show
-        render plain: CalendarRenderer.for(current_component, params[:filter]), content_type: "type/calendar"
+        render plain: CalendarRenderer.for(current_component, filter_params), content_type: "type/calendar"
       end
 
       def meeting_calendar
