+++ decidim-meetings/app/controllers/decidim/meetings/directory/meetings_controller.rb
@@ -20,7 +20,7 @@ module Decidim
         helper_method :meetings, :search
 
         def calendar
-          render plain: CalendarRenderer.for(current_organization, params[:filter]), content_type: "type/calendar"
+          render plain: CalendarRenderer.for(current_organization, filter_params), content_type: "type/calendar"
         end
 
         private
